+++ pkg/prefetch/manager.go
@@ -43,15 +43,13 @@ func (m *CredentialManager) fetchCredentials(ctx context.Context, pod *v1.Pod) {
 	}
 
 	role := k8s.PodRole(pod)
-	resolvedRole, err := m.arnResolver.Resolve(role)
-	if err != nil {
-		return
-	}
+	sessionName := k8s.PodSessionName(pod)
+	externalID := k8s.PodExternalID(pod)
 
-	identity := &sts.RoleIdentity{
-		Role:        *resolvedRole,
-		SessionName: k8s.PodSessionName(pod),
-		ExternalID:  k8s.PodExternalID(pod),
+	identity, err := sts.NewRoleIdentity(m.arnResolver, role, sessionName, externalID)
+	if err != nil {
+		logger.Errorf("error creating role identity: %s", err.Error())
+		return
 	}
 
 	issued, err := m.fetchCredentialsFromCache(ctx, identity)
